- add `PostLayout` component for better code decomp - fix gremlin - create `types.ts` file + migrate `PostType` to there to avoid "module not found" error; it was occurring bc `PostType` was defined in `lib/posts.ts` and was referenced outside of `getStaticProps`
