+++ src/types.ts
@@ -1,0 +1,4 @@
+export enum PostType {
+  Blog,
+  BIP
+}
